Initial version of C homework example
The C homework example includes a Makefile and Cython wrapper for the student.
The example contains the usual C library directories alongside the wrapper
directory. This format is designed so that Python's unittest module can still be
used to grade homework.

In the future the student may be asked to provide their own Makefile and Cython
wrapper.